requirements.txt
whisperx
torch
torchaudio
pyannote.audio
transformers
accelerate
datasets
librosa
soundfile
fastapi
python-multipart
uvicorn
dotenv
pydub
#google-generativeai -- протоколирование через Gemini (Ai studio lab)
